requirements.txt
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.4
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
uritemplate==4.1.1
urllib3==2.0.7
